requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
configparser==7.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.10
langchain_google_genai==2.0.10
langchain_community==0.3.27
langgraph==0.6.6
llama-index-core==0.13.1
llama-index-embeddings-google-genai==0.3.0
llama-index-embeddings-huggingface==0.6.0
llama-index-llms-google-genai==0.3.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.25.1
pyparsing==3.1.4
requests==2.32.3
rsa==4.9
torch==2.7.1
tqdm==4.66.5
transformers==4.53.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
xxhash==3.5.0
